+++ src/components/DatePicker/DatePickers/DatePicker.example.js
@@ -35,6 +35,7 @@ export const Dropdown_DateRangePicker = {
         ? false
         : start.getTime() < disabledDay.getTime() && disabledDay.getTime() < end.getTime();
     },
+    hasFocusOnMount: false,
   },
   group: 'inputs',
 };
@@ -62,6 +63,7 @@ export const Dropdown_SingleDatePicker = {
         ? false
         : start.getTime() < disabledDay.getTime() && disabledDay.getTime() < end.getTime();
     },
+    hasFocusOnMount: false,
   },
   group: 'inputs',
 };
@@ -97,6 +99,7 @@ export const Plain_SingleDate = {
         ? false
         : start.getTime() < disabledDay.getTime() && disabledDay.getTime() < end.getTime();
     },
+    hasFocusOnMount: false,
   },
   group: 'inputs',
 };
@@ -131,6 +134,7 @@ export const Plain_DateRange = {
         ? false
         : start.getTime() < disabledDay.getTime() && disabledDay.getTime() < end.getTime();
     },
+    hasFocusOnMount: false,
   },
   group: 'inputs',
 };